Extract integer parsing into new class Integer
